requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
sqlparse==0.2.4
djongo
flake8>=3.6.0,<3.7.0
